requirements.txt
parse
requests
colorama 
fake_useragents
